Create and migrate to CommonEconomy (#412)
* Start CommonEconomy implementation

* Stuff

* More common economy

* More CommonEconomy stuff

* Checkpoint

* More migrations to TransactionResult

* Implement more CommonEconomy

* Fix some tests

* Add a check for the toUuid in transfer function

* Modify balance command to use CommonEconomy

* Update balance commands to use CommonEconomy

* Remove CommonEconomy interface

* Migrate Bukkit EconomyImpl to use CommonEconomy

Still need to update all the commands to use CommonEconomy instead of EconomyImpl, though that'll be taken care of next.

* Fix some checkstyle warnings

* Checkpoint

* Lots of changes

* Add withdraw

* More CommonEconomy impl

* Update BalanceData to use BigDecimal instead of double

* Add more TODOs for the next release

+++ totaleconomy-bukkit/src/main/java/com/ericgrandt/totaleconomy/services/BalanceService.java
@@ -24,7 +24,7 @@ public class BalanceService {
             return new TransferResult(ResultType.FAILURE, "Insufficient funds");
         }
 
-        balanceData.transfer(fromPlayer, toPlayer, 1, amount);
+        balanceData.transfer(fromPlayer, toPlayer, 1, BigDecimal.valueOf(amount));
 
         return new TransferResult(ResultType.SUCCESS, "");
     }
